Extract BaseCDMSRestApiTestCase
